Refactor UMLDiagram Classes, so both Package and Class Diagrams both get handled by the same Diagram
