+++ frontend/vitruvius-editor/src/visualisation/uml/DiagramComponents.tsx
