+++ frontend/vitruvius-editor/src/visualisation/uml/DiagramContent.ts
