--- frontend/vitruvius-editor/src/visualisation/uml/UMLClassDiagram/ClassDiagramExtractor.ts
@@ -1,9 +1,0 @@
-import { Content } from "../../../model/Content";
-import { Extractor } from "../../Extractor";
-import { VisualisationWidget } from "../../VisualisationWidget";
-
-export class ClassDiagramExtractor implements Extractor {
-  extractContent(widget: VisualisationWidget<any>): Promise<Content> {
-    throw new Error("Method not implemented.");
-  }
-}
